chore(tests): remove redundant tests/criu_e2e.rs test file
The file `criu_e2e.rs` was a partial duplicate of `e2e_criu.rs`.
This was likely introduced during a branch merge.
